Refactor SVG path parsing
